+++ crates/blockchain/constants.rs
@@ -24,15 +24,15 @@ pub const TX_DATA_NON_ZERO_GAS: u64 = 68;
 // === EIP-170 constants ===
 
 // Max bytecode size
-pub const MAX_CODE_SIZE: usize = 0x6000;
+pub const MAX_CODE_SIZE: u32 = 0x6000;
 
 // === EIP-3860 constants ===
 
 // Max contract creation bytecode size
-pub const MAX_INITCODE_SIZE: usize = 2 * MAX_CODE_SIZE;
+pub const MAX_INITCODE_SIZE: u32 = 2 * MAX_CODE_SIZE;
 
 // Max non-contract creation bytecode size
-pub const MAX_TRANSACTION_DATA_SIZE: usize = 4 * 32 * 1024; // 128 Kb
+pub const MAX_TRANSACTION_DATA_SIZE: u32 = 4 * 32 * 1024; // 128 Kb
 
 // === EIP-2028 constants ===
